+++ CHANGELOG.md
@@ -25,6 +25,7 @@ details about the cause of the failure
 
 -    Fix incorrect dereference of wrapper object in `tp_repr`, which may result in a program crash
 -    Fix incorrect dereference in params array handling
+-    Fixes issue with function resolution when calling overloaded function with keyword arguments from python ([#1097][i1097])
 -    Fix `object[]` parameters taking precedence when should not in overload resolution
 -    Fixed a bug where all .NET class instances were considered Iterable
 -    Fix incorrect choice of method to invoke when using keyword arguments.
@@ -32,6 +33,11 @@ details about the cause of the failure
 -    Indexers can now be used with interface objects
 -    Fixed a bug where indexers could not be used if they were inherited
 -    Made it possible to use `__len__` also on `ICollection<>` interface objects
+-    Made it possible to call `ToString`, `GetHashCode`, and `GetType` on inteface objects
+
+### Removed
+
+-   implicit assembly loading (you have to explicitly `clr.AddReference` before doing import)
 
 ## [2.5.0][] - 2020-06-14